Refactor mood modules to domain directory
Moved mood-related modules from 'mood' to 'domain' directory and updated all import paths accordingly. Also updated exports in ai/prompts/index.ts for better modularity.
