+++ src/app/components/TomieTerminal/hooks/useTerminalSetup.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState, useMemo } from 'react';
-import { Mood } from '../../../mood/moodConfig';
+import { Mood } from '../../../domain/moodConfig';
 import { useI18n } from '../../../i18n/useI18n';
 import { Message } from '../../../types';
