+++ src/app/components/TomieTerminal/utils/typingEffect.ts
@@ -1,4 +1,4 @@
-import { Mood } from '../../../mood/moodConfig';
+import { Mood } from '../../../domain/moodConfig';
 import { Message } from '../../../types';
 
 export const typeMessage = async (
