+++ src/app/domain/moodConfig.ts
