+++ src/app/domain/stateManager.ts
