+++ src/app/services/ai/prompts/index.ts
@@ -1,2 +1,4 @@
-export { predefinedResponses } from './responses';
-export * from './prompts';
+export * from './creatorInfo';
+export * from './moodPersonalities';
+export * from './privacyPrompts';
+export * from './systemPrompts';
